Added Penguins image as another example (bigger).
Fixed schemna so JFIF and JIF file formats are accepted.

Added unit test for IMG_2308 used by our perf tests.

So the only actual change is that the SOI and EOI segments are now
explicitly required at the start and end of file. The rest of the
segments are all optional, or can be datablobs.

This schema requires a rich schematron validation to be used after
parsing if you want to reject non-well-formed JPEG files. Parsing with
this schema is much too mermissive.